create domain modals and mappers

+++ app/src/main/java/com/aminovic/amadeustest/data/remote/dto/WeatherDataDto.kt
@@ -7,7 +7,7 @@ data class WeatherDataDto(
     var cityDto: CityDto? = CityDto(),
 
     @field:Json(name = "time")
-    var time: Int? = null,
+    var time: Long? = null,
 
     @field:Json(name = "main")
     var main: MainDto? = MainDto(),
